requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==41.0.1
defusedxml==0.7.1
dj-database-url==2.0.0
dj-email-url==1.0.6
Django==4.2.3
django-allauth==0.54.0
django-cache-url==3.4.4
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
environs==9.5.0
gunicorn==20.1.0
idna==3.4
marshmallow==3.19.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
whitenoise==6.5.0
